Move DistributionEvent to impl
